==> MI/00.cellstate_markers.ipynb <==
#!/usr/bin/env python
# coding: utf-8
import sys
import scanpy as sc
import anndata
import pandas as pd
import numpy as np
import os
import gc
import decoupler as dc
import regex as re
import math
os.chdir("/Users/ricardoramirez/Dropbox/PhD/Research/MOFAcell")

sc_dat = sc.read_h5ad(filename = "/Volumes/RicoData2/MI_project/MI_revisions/HCA_submission/snrnaseq/hca/hca_rnasamples.h5ad")

sc_dat = sc_dat[~sc_dat.obs["cell_state"].isna().values]

# Next filter out cells that are part of analyzed cells
cts = ["CM","Fib","Endo","Myeloid"]
degs = {}

for ct in cts:
    print(ct)
    ct_dat = sc_dat[sc_dat.obs["cell_type"].values == ct].copy()
    sc.pp.normalize_total(ct_dat, target_sum=1e4)
    sc.pp.log1p(ct_dat)
    ct_dat.raw = ct_dat
    sc.tl.rank_genes_groups(ct_dat, groupby="cell_state",use_raw=True)
    mrkrs = sc.get.rank_genes_groups_df(ct_dat, group = None)
    mrkrs["ct"] = ct
    degs[ct] = mrkrs

deg_df = pd.concat(degs)
deg_df.to_csv("./data_MI/cell_states_mrkrs_ttest.csv")

deg_df
