==> MI/00.createstatepb.ipynb <==
#!/usr/bin/env python
# coding: utf-8
import sys
import scanpy as sc
import anndata
import pandas as pd
import numpy as np
import os
import gc
import decoupler as dc
import regex as re
import math
os.chdir("/Users/ricardoramirez/Dropbox/PhD/Research/MOFAcell")

sc_dat = sc.read_h5ad(filename = "/Volumes/RicoData2/MI_project/MI_revisions/HCA_submission/snrnaseq/hca/hca_rnasamples.h5ad")

sc_dat = sc_dat[~sc_dat.obs["cell_state"].isna().values]

sc_dat.layers['counts'] = sc_dat.X

cell_type_numbers = sc_dat.obs.groupby(["patient_region_id", "cell_state"])["cell_state"].count()
cell_type_numbers = cell_type_numbers.to_frame().rename({"cell_state":"counts"}, axis=1).reset_index()

padata = dc.get_pseudobulk(sc_dat, sample_col='patient_region_id', groups_col='cell_state', layer='counts', min_prop=0, min_smpls=0)

pb_dat = pd.DataFrame(padata.X)
pb_dat.columns = padata.var.index.values
pb_dat.index = padata.obs.index.values
pb_dat.to_csv("./data_MI/pb_data_states.csv")

pb_coldata = padata.obs.copy()
pb_coldata["colname"] = pb_coldata.index.values
pb_coldata = pb_coldata.merge(cell_type_numbers, on = ["cell_state","patient_region_id"], how = "left")

pb_coldata.to_csv("./data_MI/pb_coldata_states.csv")

pb_coldata
