==> MI/02.2.stateenrichment.Rmd <==
---
title: "02.2.stateenrichment"
author: "Ricardo Ramirez"
date: "2023-01-17"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Introduction

One possibility is that by ignoring individual cells, MOFAcell can't capture cell diversity, nevertheless I argue that the genes associated with the loadings can capture the changes in compositions in cell-states associated with a certain covariate

```{r setup, include=FALSE}
library(tidyverse)
library(cowplot)
source("MOFAcell/MOFAcellprep.R")
```

# Import composition info of cell states

```{r}
annotation_names <- tibble(patient_group = c("group_1", "group_2", "group_3"),
                           patient_group_name = c("myogenic", "ischemic", "fibrotic"))


state_coldata <- read_csv("./data_MI/pb_coldata_states.csv",show_col_types = F)[, -1] %>%
  column_to_rownames("colname") %>%
  dplyr::select(cell_type, cell_state, patient_region_id, patient_group, counts) %>%
  dplyr::rename("donor_id" = "patient_region_id",
                "cell_counts" = "counts") %>%
  group_by(donor_id, cell_type) %>%
  dplyr::mutate(donor_cell_counts = sum(cell_counts)) %>%
  dplyr::mutate(cell_props = cell_counts/donor_cell_counts) %>%
  left_join(annotation_names)  %>%
  dplyr::select(-patient_group) %>%
  dplyr::rename("patient_group" = patient_group_name) %>%
  dplyr::mutate(patient_group = factor(patient_group, 
                                       levels = c("myogenic", "fibrotic", "ischemic")))
```

# Import loadings

```{r}
# In this function we separate loading genes from
# positive and negative loadings, and estimate celltypes

clean_loads <- function(loadings, sep = "_") {
  
  c_load <- loadings %>%
    dplyr::mutate(ct = strsplit(celltype,"_") %>%
                    map_chr(., ~.x[1]),
                  class = strsplit(celltype,"_") %>%
                    map_chr(., ~.x[2])) %>%
    dplyr::select(gene, ct, class)
  
  return(c_load)
  
}

```

```{r}
cts <- c("CM", "Endo", "Myeloid", "Fib")

loadings <- read_csv("./results/MI/MOFA_mcell/factor_desc/Factor1_char/loadings.csv",
                     show_col_types = F) %>%
  clean_loads(.) %>%
  dplyr::rename("view" = "ct",
                "feature" = "gene")

head(loadings)
```

# First enrich cell-state markers into the loading matrix

Here we use overrepresentation analysis with hypergeometric tests

```{r}
GSE_analysis = function(geneList,Annotation_DB){
  
  geneList = geneList[geneList %in% unique(unlist(Annotation_DB))]
  
  ResultsDF = matrix(0,nrow = length(Annotation_DB),ncol = 5)
  rownames(ResultsDF) = names(Annotation_DB)
  colnames(ResultsDF) = c("GenesInPathway","GenesInList","GeneNames","p_value","corr_p_value")
  
  DB_genecontent = length(unique(unlist(Annotation_DB)))
  GenesDB = DB_genecontent 
  SelectedGenes = length(geneList)
  
  for(gset in rownames(ResultsDF)){
    GP = length(Annotation_DB[[gset]])
    GL = length(intersect(Annotation_DB[[gset]],geneList))
    
    ResultsDF[gset,"GenesInList"] = GL
    ResultsDF[gset,"GenesInPathway"] = GP
    ResultsDF[gset,"GeneNames"] = paste(intersect(Annotation_DB[[gset]],geneList),collapse = ",")
    ResultsDF[gset,"p_value"] = phyper(q=GL - 1, m=GP, n=GenesDB-GP, k=SelectedGenes, lower.tail = FALSE, log.p = FALSE)
  }
  
  ResultsDF[,"corr_p_value"] = p.adjust(ResultsDF[,"p_value"],method = "BH")
  ResultsDF = data.frame(ResultsDF,stringsAsFactors = F)
  ResultsDF = ResultsDF[order(ResultsDF[,"p_value"]),]
  
  ResultsDF = ResultsDF %>% 
    rownames_to_column("gset") %>% 
    mutate_at(c("GenesInPathway","GenesInList",
                "p_value","corr_p_value"), 
              as.numeric) %>% 
    dplyr::arrange(corr_p_value,GenesInList)
  
  return(ResultsDF)
}
```

```{r}
# State markers from the Kuppe et. al. paper, recalculated with t-tests
state_mrkrs <- read_csv("./data_MI/cell_states_mrkrs_ttest.csv", show_col_types = F) %>%
  dplyr::filter(pvals_adj < 0.05, logfoldchanges >= 0.5) %>%
  dplyr::mutate(cs_name = paste0(ct, "_", group)) %>%
  dplyr::select(cs_name, names) %>%
  group_by(cs_name) %>%
  nest() %>%
  dplyr::mutate(data = map(data, ~.x[[1]])) %>%
  deframe()

# Enriching states into the loadings
enrichment_per_ct <- set_names(cts) %>%
  map(., function(ct) {
    
    useful_states <- grepl(ct, names(state_mrkrs))
    useful_states <- state_mrkrs[useful_states]
    
    enrichment <- dplyr::filter(loadings, view == ct) %>%
      group_by(view,class) %>%
      nest() %>%
      dplyr::mutate(data = map(data, ~.x[[1]])) %>%
      dplyr::mutate(enrichment = map(data, GSE_analysis, Annotation_DB = useful_states)) %>%
      dplyr::select(view,class, enrichment) %>%
      unnest(cols = c(enrichment)) %>%
      dplyr::select(view, class, gset, corr_p_value)
    
    return(enrichment)
  })

# Making plots per cell type
plt_per_ct <- map(enrichment_per_ct, function(dat) {
  
  df_plt <- dat %>%
    dplyr::mutate(class = ifelse(class == "neg", "healthy", "disease") %>%
                    factor(., levels = c("healthy", "disease"))) %>%
    dplyr::mutate(MOFA_loading = paste0(view,"_",class),
                  logpval = -log10(corr_p_value)) %>%
    dplyr::rename("state" = gset)
  
  
  enrich_plt <- ggplot(df_plt, aes(y = state, x = MOFA_loading, size = logpval)) +
    geom_point() +
    theme_minimal() +
    theme(axis.text.x = element_text(size = 11, angle = 90, hjust = 0, vjust = 0.5),
          axis.text.y = element_text(size = 11)) +
    ggtitle("enrichment") +
    coord_equal()
  
  return(enrich_plt)
  
})
```

# Plots for supplement

```{r, fig.height=3, fig.width= 5}
pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/state_relationship.pdf", height = 2.5, width = 4.5)
walk(plt_per_ct, plot)
dev.off()
```

# Plots for main

We will pick the most significant states per direction

```{r}
healthy_sets <- map(enrichment_per_ct, function(ct_dat) {
  ct_dat %>%
    filter(corr_p_value == min(corr_p_value),
           class == "neg") %>%
    pull(gset)
})

disease_sets <- map(enrichment_per_ct, function(ct_dat) {
  ct_dat %>%
    filter(corr_p_value == min(corr_p_value),
           class == "pos") %>%
    pull(gset)
})
```

Here we make the joint barplots to show how state enrichment and median compositions align

```{r}
state_plt_main <- function(ct, state_sets, color_fill) {
  
  enrich_dat <- enrichment_per_ct[[ct]] %>%
      dplyr::filter(gset %in% state_sets[[ct]]) %>%
    dplyr::mutate(class = ifelse(class == "neg", "healthy", "disease") %>%
                    factor(., levels = c("healthy", "disease"))) %>%
    dplyr::mutate(MOFA_loading = paste0(view,"_",class),
                  logpval = -log10(corr_p_value))
    
    enrich_bar <- ggplot(enrich_dat, aes(x = class, y = logpval)) +
      geom_bar(stat = "identity", fill = color_fill) +
      theme_classic() +
      theme(axis.text = element_text(size = 10),
            axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
            axis.title = element_text(size = 10),
            plot.title = element_text(size = 8, face = "bold")) +
      ylab("-log10(pval)") +
      xlab("") +
      ggtitle(gsub(paste0("^", ct, "_"), "", state_sets[[ct]]))

    props_dat <- state_coldata %>%
      dplyr::filter(cell_type == ct,
                    cell_state == gsub(paste0("^", ct, "_"), "", state_sets[[ct]])) %>%
      group_by(patient_group, cell_state) %>%
      dplyr::summarise(mean_prop = mean(cell_props),
                       sd_prop = sd(cell_props))
    
    comp_bar <- ggplot(props_dat, aes(x = patient_group, y = mean_prop)) +
      geom_bar(stat = "identity", fill = color_fill) +
      geom_errorbar(aes(x=patient_group, 
                        ymin=mean_prop-sd_prop, 
                        ymax=mean_prop+sd_prop), 
                    width=0.2, colour="black", alpha=0.7, size=0.5)+
      theme_classic() +
      theme(axis.text = element_text(size = 10),
            axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
            axis.title = element_text(size = 10),
            plot.title = element_text(size = 8, face = "bold")) +
      ylab("mean prop.") +
      xlab("") +
      ggtitle(gsub(paste0("^", ct, "_"), "", state_sets[[ct]]))
    
    all_plts <- cowplot::plot_grid(enrich_bar, comp_bar, align = "hv", ncol = 2,rel_widths = c(0.8, 1))
    
    return(all_plts)

}
```

```{r}
state_coldata %>% 
  group_by(cell_state, patient_group) %>% 
  summarize(n()) %>% 
  dplyr::filter(cell_state %in% c("Capillary_Endo", "LYVE_FOLR_Macrophages","healthy_CM", "Fib_0"))


state_coldata %>% 
  group_by(cell_state, patient_group) %>% 
  summarize(n()) %>% 
  dplyr::filter(cell_state %in% c("damaged_CM", "Lymphatic_Endo", "SPP1_Macrophages", "Myofib"))

```


```{r}
healthy_plts <- set_names(cts) %>%
  map(., state_plt_main, state_sets = healthy_sets, color_fill = "#4169E1")

disease_plts <- set_names(cts) %>%
  map(., state_plt_main, state_sets = disease_sets, color_fill = "#FF6666")

ct_plts <- set_names(cts) %>%
  map(., function(ct) {
    
    j_plot <- cowplot::plot_grid(healthy_plts[[ct]], 
                                disease_plts[[ct]],
                                ncol = 2,
                                align = "hv")
    
    return(j_plot)
    
  })

pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/all_states_bars.pdf", height = 8, width = 5)
cowplot::plot_grid(plotlist = ct_plts, ncol = 1, nrow = 4, align = "hv")
dev.off()
```
