==> MI/02.3.funceval.Rmd <==
---
title: "Functional Enrichment"
author: "Ricardo Ramirez"
date: "11/2/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Functional characterization of gene loadings

```{r setup, include=FALSE}
library(tidyverse)
library(ComplexHeatmap)
library(decoupleR)
source("./MOFAcell/factorutils.R")
```

```{r}
loadings <- read_csv("./results/MI/MOFA_mcell/factor_desc/Factor1_char/loadings.csv") %>%
  dplyr::select(-value) %>%
  dplyr::group_by(celltype) %>%
  nest() %>%
  dplyr::mutate(data = map(data, ~.x[[1]])) %>%
  deframe()
```

```{r}
gsets <- readRDS("./aux_data/Genesets_Dec19.rds")
```

## Hypergeometric test for functional enrichment

```{r}

GSE_analysis <- function(geneList,Annotation_DB){
  
  geneList = geneList[geneList %in% unique(unlist(Annotation_DB))]
  
  ResultsDF = matrix(0,nrow = length(Annotation_DB),ncol = 5)
  rownames(ResultsDF) = names(Annotation_DB)
  colnames(ResultsDF) = c("GenesInPathway","GenesInList","GeneNames","p_value","corr_p_value")
  
  DB_genecontent = length(unique(unlist(Annotation_DB)))
  GenesDB = DB_genecontent 
  SelectedGenes = length(geneList)
  
  for(gset in rownames(ResultsDF)){
    GP = length(Annotation_DB[[gset]])
    GL = length(intersect(Annotation_DB[[gset]],geneList))
    
    ResultsDF[gset,"GenesInList"] = GL
    ResultsDF[gset,"GenesInPathway"] = GP
    ResultsDF[gset,"GeneNames"] = paste(intersect(Annotation_DB[[gset]],geneList),collapse = ",")
    ResultsDF[gset,"p_value"] = phyper(q=GL - 1, m=GP, n=GenesDB-GP, k=SelectedGenes, lower.tail = FALSE, log.p = FALSE)
  }
  
  ResultsDF[,"corr_p_value"] = p.adjust(ResultsDF[,"p_value"],method = "BH")
  ResultsDF = data.frame(ResultsDF,stringsAsFactors = F)
  ResultsDF = ResultsDF[order(ResultsDF[,"p_value"]),]
  
  ResultsDF = ResultsDF %>% 
    rownames_to_column("gset") %>% 
    mutate_at(c("GenesInPathway","GenesInList",
                "p_value","corr_p_value"), 
              as.numeric) %>% 
    dplyr::arrange(corr_p_value,GenesInList)
  
  return(ResultsDF)
}
```

# Main

```{r}
cts_enrichment <- map(loadings, GSE_analysis, Annotation_DB = gsets$MSIGDB_HMARKS)
```

```{r}
col_fun_fact <- circlize::colorRamp2(seq(0, 5, length = 50), hcl.colors(50, "Blues",rev = T))

neg_sets <- cts_enrichment %>% 
  enframe() %>%
  unnest() %>%
  dplyr::filter(corr_p_value < 0.1,
                grepl("neg",name)) %>%
  dplyr::group_by(name) %>%
  pull(gset)
  
neg_enrichment <- cts_enrichment %>% 
  enframe() %>%
  unnest() %>%
  dplyr::filter(gset %in% neg_sets,
                grepl("neg",name)) %>%
  dplyr::mutate(corr_p_value = -log10(corr_p_value)) %>%
  dplyr::mutate(corr_p_value = ifelse(corr_p_value > 5, 5, corr_p_value)) %>%
  dplyr::mutate(gset = gsub("HALLMARK_", "", gset) %>% strtrim(width = 40),
                name = gsub("_neg", "", name))
                
pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/functional_healthy.pdf", height = 3.4, width = 3.8)

neg_plot <- neg_enrichment %>%
  dplyr::select(name, gset, corr_p_value) %>%
  pivot_wider(names_from = name, values_from = corr_p_value) %>%
  column_to_rownames("gset") %>%
  as.matrix() %>%
  ComplexHeatmap::Heatmap(., name = "healthy \n-log(pval)",
                          show_row_dend = F, 
                          show_column_dend = F, 
                          row_names_gp = grid::gpar(fontsize = 8),
                          column_names_gp = grid::gpar(fontsize = 10),
                          col = col_fun_fact)

plot(neg_plot)

dev.off()

```

```{r}
col_fun_fact <- circlize::colorRamp2(seq(0, 5, length = 50), hcl.colors(50,"Reds",rev = T))

pos_sets <- cts_enrichment %>% 
  enframe() %>%
  unnest() %>%
  dplyr::filter(corr_p_value <= 0.1,
                grepl("pos",name)) %>%
  dplyr::group_by(name) %>%
  dplyr::slice(1:6) %>%
  pull(gset)
  
pos_enrichment <- cts_enrichment %>% 
  enframe() %>%
  unnest() %>%
  dplyr::filter(gset %in% pos_sets,
                grepl("pos",name)) %>%
  dplyr::mutate(corr_p_value = -log10(corr_p_value)) %>%
  dplyr::mutate(corr_p_value = ifelse(corr_p_value > 5, 5, corr_p_value)) %>%
  dplyr::mutate(gset = gsub("HALLMARK_", "", gset) %>% strtrim(width = 35),
                name = gsub("_pos", "", name))

pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/functional_dis.pdf", height = 3, width = 3.8)

pos_plot <- pos_enrichment %>%
  dplyr::select(name, gset, corr_p_value) %>%
  pivot_wider(names_from = name, values_from = corr_p_value) %>%
  column_to_rownames("gset") %>%
  ComplexHeatmap::Heatmap(., name = "disease \n-log(pval)",
                          show_row_dend = F, 
                          show_column_dend = F, 
                          row_names_gp = grid::gpar(fontsize = 8),
                          column_names_gp = grid::gpar(fontsize = 10),
                          col = col_fun_fact)

plot(pos_plot)

dev.off()
```

```{r}
pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/functional_all.pdf", height = 4.8, width = 4.5)
ht_list <- neg_plot %v% pos_plot
draw(ht_list)
dev.off()
```

# An alternative analysis is to use PROGENy

```{r}
model_outfile <- "./results/MI/MOFA_mcell/MI_model6factors.hdf5"
model <- MOFA2::load_model(model_outfile)
```

```{r}
factor_loadings <- get_floadings(model = model,
                                   factor = "Factor1") %>%
  dplyr::filter(abs(value)>=0.1) %>%
  pivot_wider(names_from = ctype, values_from = value, values_fill = 0) %>%
  column_to_rownames("feature") %>%
  as.matrix()

```

```{r}
net <- get_progeny(organism = 'human', top = 1000)
```

```{r}
net <- gsets$MSIGDB_HMARKS %>%
  enframe(name = "source",value = "target") %>%
  unnest() %>% 
  dplyr::mutate(weight = 1,
                source =  gsub("HALLMARK_", "", source))
  
```

```{r}
col_fun_scores <- circlize::colorRamp2(seq(-5, 5, length = 50), hcl.colors(50,"TealRose",rev = T))

ct_acts <- run_wmean(mat=factor_loadings, net=net, .source='source', .target='target',
                  .mor='weight', times = 1000, minsize = 5)

ct_acts_mat <- ct_acts %>%
  dplyr::filter(statistic == "norm_wmean") %>%
  dplyr::mutate(score = ifelse(score <= -5, -5, score)) %>%
  dplyr::select(source, condition, score) %>%
  pivot_wider(values_from = score, names_from = condition) %>%
  column_to_rownames("source") %>%
  as.matrix()

hmap <- ComplexHeatmap::Heatmap(t(ct_acts_mat), 
                        name = "progeny", 
                        col = col_fun_scores, 
                        show_row_dend = FALSE,
                        show_column_dend = FALSE)

pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/progeny_res.pdf", height = 2.7, width = 5.5)
draw(hmap)
dev.off()

draw(hmap)
```

## Hallmarks

```{r}
net <- gsets$MSIGDB_HMARKS %>%
  enframe(name = "source",value = "target") %>%
  unnest() %>% 
  dplyr::mutate(weight = 1,
                source =  gsub("HALLMARK_", "", source))

ct_acts <- run_wmean(mat=factor_loadings, net=net, .source='source', .target='target',
                  .mor='weight', times = 1000, minsize = 5)

```

```{r}
col_fun_scores <- circlize::colorRamp2(seq(-5, 5, length = 50), hcl.colors(50,"Blue-Red 2",rev = F))

ct_acts_mat <- ct_acts %>%
  dplyr::filter(statistic == "norm_wmean") %>%
  dplyr::mutate(score = ifelse(score <= -5, -5, score)) %>%
  dplyr::select(source, condition, score) %>%
  pivot_wider(values_from = score, names_from = condition) %>%
  column_to_rownames("source") %>%
  as.matrix()

sel_paths <- names(rowMeans(abs(ct_acts_mat)) %>% sort(decreasing = TRUE)) 


hmap <- ComplexHeatmap::Heatmap((ct_acts_mat[sel_paths[1:25],]), 
                        name = "normalized \nscore", 
                        col = col_fun_scores, 
                        show_row_dend = FALSE,
                        row_names_gp = grid::gpar(fontsize = 8),
                        column_names_gp = grid::gpar(fontsize = 11),
                        show_column_dend = FALSE)

pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/hallmarks_res.pdf", height = 3.8, width = 4.3)
draw(hmap)
dev.off()

draw(hmap)
```
