==> MI/04.01.calculateeffectiveareas_ct.Rmd <==
---
title: "Area estimation ct based"
author: "Ricardo Ramirez"
date: "7/29/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
library(tidyverse)
```

## Defining effective areas

We create regions of interest for each cell-type by identifying spots where msk_thrsh is greater than a certain proportion, this ensures that any module score calculated before is reliable

Then from this mask, we calculate potential effective areas and real effective areas with sd_thrsh (since we are doing weighted means)

```{r}
process_modules <- function(module_file, 
                            sd_thrsh = 2,
                            msk_thrsh = 0.1) {
  
  # This recovers the module scores
  module_mat <- read_csv(module_file,show_col_types = F) %>%
    as.data.frame()
  
  fit_module_file <- gsub("[.]csv", "_msk.csv", module_file)
  
  rownames(module_mat) <- module_mat[,1]
  module_mat <- module_mat[, -1]
  
  # Filter compositions
  slide_comps <- compositions[rownames(module_mat),]
  
  # Mask compositions
  slide_comps[slide_comps < msk_thrsh] = 0
  slide_comps[slide_comps != 0] = 1
  
  # Get ct order
  ct_order <- colnames(module_mat) %>% 
  strsplit(., "_") %>%
  map_chr(., ~.x[[1]])
  
  slide_comps <- slide_comps[rownames(module_mat), ct_order]
  colnames(slide_comps) <- colnames(module_mat)
  
  # Get masked module score matrix
  mask_module_mat <- module_mat * slide_comps
  
  mask_module_mat[mask_module_mat < sd_thrsh] <- 0
  
  write_csv(mask_module_mat %>% 
              as.data.frame() %>%
              rownames_to_column("spot_id"),
            fit_module_file)
  
  module_area <- slide_comps %>%
  colSums() %>%
  enframe(value = "module_totaln")
  
  summ <- mask_module_mat %>%
    as.data.frame() %>%
    rownames_to_column("spot_id") %>%
    pivot_longer(-spot_id) %>%
    group_by(name) %>%
    mutate(n_module = map_lgl(value, ~ .x != 0) %>%
             sum()) %>%
    left_join(module_area, by = "name") %>%
    dplyr::filter(value != 0) %>%
    mutate(mean_module = mean(value)) %>%
    dplyr::select(name, n_module, mean_module, module_totaln) %>%
    unique() %>%
    mutate(area = n_module/module_totaln,
           w_area = (n_module/module_totaln) * mean_module)
  
  return(summ)
  
  }
```

# Main

## Define composition matrix

```{r}
compositions <- read_csv("./data_MI/cell_type_compositions.csv")

compositions <- compositions %>%
  dplyr::mutate(spot_id = gsub("[..]","_", spot_id)) %>%
  dplyr::mutate(spot_id = gsub("__","_", spot_id))

compositions <- compositions %>%
  dplyr::select(spot_id, name, value) %>%
  pivot_wider(names_from = name,values_from = value) %>%
  column_to_rownames("spot_id") %>%
  as.matrix()
```

## Define meta data

```{r}
annotation_names <- tibble(patient_group = c("group_1", 
                                             "group_2", 
                                             "group_3"),
                           patient_group_name = factor(c("myogenic", 
                                                  "ischemic", 
                                                  "fibrotic"),
                                                  levels = c("myogenic", 
                                                  "ischemic", 
                                                  "fibrotic")))

meta <- read_csv("./data_MI/visium_patient_anns_revisions.csv") %>%
  left_join(annotation_names)
 
```

# Factor 1

```{r, warning=FALSE, message=FALSE}
dir <- "./results/MI/MOFA_mcell/factor_desc/Factor1_char/decoupler_ct/"

samples <- list.files(dir)
samples <- samples[!grepl("msk",samples)]

samples <- set_names(paste0(dir,samples), gsub("[.]csv", "", samples))

area_summ <- map(samples, process_modules, sd_thrsh = 2, msk_thrsh = 0.1) %>%
  enframe(name = "sample_id") %>%
  unnest() %>%
  dplyr::select(sample_id, name, area, w_area, mean_module) %>%
  tidyr::complete(sample_id, name, fill = list("area" = 0,
                                               "w_area" = 0,
                                               "mean_module" = 0)) %>%
  left_join(meta, by = "sample_id") %>%
  dplyr::mutate(name = ifelse(grepl("_neg", name), 
                              gsub("_neg", "_healthy", name),
                              gsub("_pos", "_disease", name)))

write.csv(area_summ, "./results/MI/MOFA_mcell/factor_desc/Factor1_char/spatial_summ.csv")

area_summ %>% group_by(name, patient_group_name) %>% summarize(n()) %>% print(n = 50)
```

## Areas with "cell-type" program

```{r}
stat.test <- area_summ %>%
  ggpubr::compare_means(formula = area ~ patient_group_name, 
                        group.by = "name")

stat.test %>%
  dplyr::filter(p.adj <= 0.1)
```

```{r, fig.width=5, height = 9}
plt_sum <- area_summ %>%
  ggplot(aes(x = patient_group_name, y = area)) +
  geom_boxplot() +
  theme_classic() +
  facet_wrap( ~ name, ncol = 4, scales = "free") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) 

plot(plt_sum)

pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/area_spatial.pdf", height = 6.5, width = 5)

plot(plt_sum)

dev.off()
```

## Mean "cell-type" program expression

```{r}
stat.test = area_summ %>%
  ggpubr::compare_means(formula = mean_module ~ patient_group_name, 
                        group.by = "name")

stat.test %>%
  dplyr::filter(p.adj < 0.05)
```


```{r, fig.width=5, height = 9}
plt_sum <- area_summ %>%
  ggplot(aes(x = patient_group_name, y = mean_module)) +
  geom_boxplot() +
  theme_classic() +
  facet_wrap( ~ name, ncol = 4, scales = "free") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) 

plot(plt_sum)

pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/moduleexpr_spatial.pdf", height = 6.5, width = 4.5)

plot(plt_sum)

dev.off()


```

## Mean "cell-type" program expression weighted by area

```{r}
stat.test = area_summ %>%
  ggpubr::compare_means(formula = w_area ~ patient_group_name, 
                        group.by = "name", method = "wilcox.test")

stat.test %>%
  dplyr::filter(p.adj < 0.05)
```

```{r}
stat.test %>%
  dplyr::filter(p.adj < 0.1)
```


```{r, fig.width=5, height = 9}
plt_sum <- area_summ %>%
  ggplot(aes(x = patient_group_name, y = mean_module)) +
  geom_boxplot() +
  theme_classic() +
  facet_wrap( ~ name, ncol = 4, scales = "free") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) 

plot(plt_sum)

pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/area_wmoduleexpr_spatial.pdf", height = 6.5, width = 4.5)

plot(plt_sum)

dev.off()


```



```{r, fig.width=5, height = 7}
max_val <- max(area_summ$w_area) + 0.05

disease_area <- area_summ %>%
  dplyr::mutate(ct = strsplit(name, "_") %>%
                  map_chr(., ~.x[[1]]),
                direction = strsplit(name, "_") %>%
                  map_chr(., ~.x[[2]])) %>%
  dplyr::filter(direction == "disease") %>%
  dplyr::filter(ct %in% c("CM","Fib","Endo","Myeloid")) %>%
  ggplot(aes(x = patient_group_name, y = w_area, color = patient_group_name)) +
  geom_boxplot() +
  geom_point(alpha = 0.7) +
  facet_grid(ct ~ direction, scales = "free_y") +
  theme_minimal() +
  theme(axis.text.x = element_text(size= 11, angle = 90, hjust =1, vjust = 0.5)) +
  theme(axis.text.y = element_text(size = 11)) +
  theme(panel.spacing = unit(2, "lines")) +
  ylab("Weighted activation area") +
  xlab("")


pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/relative_area_disease.pdf", height = 6, width = 1.5)

plot(disease_area)

dev.off()
```

```{r}
max_val <- max(area_summ$w_area) + 0.05

healthy_area <- area_summ %>%
  dplyr::mutate(ct = strsplit(name, "_") %>%
                  map_chr(., ~.x[[1]]),
                direction = strsplit(name, "_") %>%
                  map_chr(., ~.x[[2]])) %>%
  dplyr::filter(direction == "healthy") %>%
  dplyr::filter(ct %in% c("CM","Fib","Endo","Myeloid")) %>%
  ggplot(aes(x = patient_group_name, y = w_area, color = patient_group_name)) +
  geom_boxplot() +
  geom_point(alpha = 0.7) +
  facet_grid(ct ~ direction, scales = "free_y") +
  theme_minimal() +
  theme(axis.text.x = element_text(size= 11, angle = 90, hjust =1, vjust = 0.5)) +
  theme(axis.text.y = element_text(size = 11)) +
  theme(panel.spacing = unit(2, "lines")) +
  ylab("Weighted activation area") +
  xlab("")


pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/relative_area_healthy.pdf", height = 6, width = 1.5)

plot(healthy_area)

dev.off()
```

```{r}
pdf("./results/MI/MOFA_mcell/factor_desc/Factor1_char/relative_area_all.pdf", height = 6, width = 6)
plot(cowplot::plot_grid(disease_area, healthy_area, ncol = 2, align = "hv"))
dev.off()

```

# Selecting contrasting results

```{r}
area_summ %>%
  dplyr::mutate(ct = strsplit(name, "_") %>%
                  map_chr(., ~.x[[1]]),
                direction = strsplit(name, "_") %>%
                  map_chr(., ~.x[[2]])) %>%
  dplyr::filter(direction == "disease") %>%
  dplyr::filter(ct %in% c("CM","Fib","Endo","Myeloid")) %>%
  group_by(ct) %>%
  dplyr::mutate(rank_expr = rank(w_area)) %>%
  group_by(sample_id, patient_group_name) %>%
  dplyr::summarise(gen_rank = mean(rank_expr)) %>%
  arrange(patient_group_name, gen_rank) %>%
  group_by(patient_group_name) %>%
  dplyr::slice(1:2)
```

```{r}
area_summ %>%
  dplyr::mutate(ct = strsplit(name, "_") %>%
                  map_chr(., ~.x[[1]]),
                direction = strsplit(name, "_") %>%
                  map_chr(., ~.x[[2]])) %>%
  dplyr::filter(direction == "disease") %>%
  dplyr::filter(ct %in% c("CM","Fib","Endo","Myeloid")) %>%
  group_by(ct) %>%
  dplyr::mutate(rank_expr = rank(w_area)) %>%
  group_by(sample_id, patient_group_name) %>%
  dplyr::summarise(gen_rank = mean(rank_expr)) %>%
  arrange(patient_group_name, desc(gen_rank)) %>%
  group_by(patient_group_name) %>%
  dplyr::slice(1:2)
```
